requirements.txt
streamlit
PyPDF2
google-genai
Pillow
pandas